+++ WorkoutDone/WorkoutDone/Sources/Classes/UIComponent/GradientButton.swift
@@ -1,0 +1,33 @@
+//
+//  GradientButton.swift
+//  WorkoutDone
+//
+//  Created by hyemi on 2023/03/04.
+//
+
+import UIKit
+
+class GradientButton: UIButton {
+    let gradient = CAGradientLayer()
+    
+    init(colors: [CGColor]) {
+        super.init(frame: .zero)
+
+        gradient.colors = colors
+        gradient.locations = [0.0 , 1.0]
+        gradient.startPoint = CGPoint(x: 0.25, y: 0.5)
+        gradient.endPoint = CGPoint(x: 0.75, y: 0.5)
+        gradient.cornerRadius = 12
+        
+        layer.addSublayer(gradient)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError()
+    }
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        gradient.frame = bounds
+    }
+}
